+++ processor.py
@@ -1295,7 +1295,7 @@ def handle_save_component(component_name, component_attribute, body, is_check_on
                     time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(time.time())))))
 
             # If succeed, just hide it in two seconds later
-            sublime.set_timeout_async(util.hide_output_panel, 2 * 1000)
+            sublime.set_timeout_async(util.hide_output_panel, 1.5 * 1000)
 
         # If not succeed, just go to the error line
         # Because error line in page is always at the line 1, so just work in class or trigger