+++ HISTORY.rst
@@ -7,7 +7,8 @@ Release History
 Release 0.7.0 (2014-08-11)
 ++++++++++++++++++
 - Even if component saving is succeed, show the output panel
-- If component saving is succeed, hide the open output panel after two seconds
+- If component saving is succeed, hide the open output panel after 1.5 seconds
+- When generating workbook or describe sobject, write the type column with Formula(<Field Type>) or <Field Type>
 
 
 Release 0.6.9 (2014-08-09)